Beginnings of the output window.
We have a function output window with input, output and interactive elements that can be customised in the future. The log file also has some updates to make sure it will always stick to the System.err and System.out even when we overwrite them with our own components later on (I think, not tested yet.).