Level and room development.
We are starting on gameplay now which is why there are so many changes. Levels, rooms and items have all been added in their most basic forms now which can then be extended or instantiated for the actual games. The main level and the bedroom have been added from the plan. The command handler has been added to allow proper interaction with the program and the basis for command hijacking has been added which we can use for some puzzles. Players have been added which just contains which room they're in and which level. It has a name which we might add later on if we want to have some sort of dialog with the user but we will see. The input area should now be in focus whenever the user opens the window which is nicer. Command defaults have been added for the most common commands to try and reduce the amount of repetative code we need to write. I am using synonyms from google right now so not all of them are perfect but I figure that it's a bit easier than making them all up myself. Any additional ones I think of can be added anyway.