Introduced graphics
There are various changes around it to make this work. The room classes now set an image to be displayed on the interactive panel when the room is entered by using the enter room function. The interactive panel now only needs to check if the image is not null as it will always draw the grey background (otherwise the background just turns white).