Level 1 is done.
That's it! Level 1 is complete! All rooms are added right now and all working. There may be a few tweaks that have to be made to descriptions and such but for now I can play through it (while knowing the commands). The alchemy puzzle is also added which makes the user pick from various coloured bottles and pots to make a solution. It is randomly generated.