Some debug fixes.
A fix to the alchemy puzzle to make it hallucinate in case you get it wrong and to remove the correct answer from it. The align puzzle now does the dripping stuff instead of pouring as it would skip over the puzzle otherwise. Some pom changes to make a working build.