Gameplay technically done.
The game now works and you can play it through (as far as testing goes). There is 1 room missing its interface which will be committed shortly. There a tonne of resource files in this one as IDEA hasn't been versioning them by default. They are all here now so you should be able to compile this directly from a clone now.