Messed up level design.
I forgot to add a room to l2 but it's there now, it just took some shuffling of files. I haven't tested this yet so there may be changes in a second. Also the metal workshop (l1r5/l2r6) has a graphic now, I just need to add frames.